+++ config/autoload.php
@@ -14,7 +14,8 @@
  */
 ClassLoader::addNamespaces(
     [
-	'HeimrichHannot',]
+        'HeimrichHannot',
+    ]
 );
 
 
@@ -23,22 +24,23 @@ ClassLoader::addNamespaces(
  */
 ClassLoader::addClasses(
     [
-	// Models
-	'HeimrichHannot\FileCredit\FileCreditModel'     => 'system/modules/filecredits/models/FileCreditModel.php',
-	'HeimrichHannot\FileCredit\FileCreditPageModel' => 'system/modules/filecredits/models/FileCreditPageModel.php',
-	'HeimrichHannot\FileCredit\FilesModel'          => 'system/modules/filecredits/models/FilesModel.php',
+        // Models
+        'HeimrichHannot\FileCredit\FileCreditModel'     => 'system/modules/filecredits/models/FileCreditModel.php',
+        'HeimrichHannot\FileCredit\FileCreditPageModel' => 'system/modules/filecredits/models/FileCreditPageModel.php',
+        'HeimrichHannot\FileCredit\FilesModel'          => 'system/modules/filecredits/models/FilesModel.php',
 
-	// Modules
-	'HeimrichHannot\FileCredit\ModuleFileCredit'    => 'system/modules/filecredits/modules/ModuleFileCredit.php',
+        // Modules
+        'HeimrichHannot\FileCredit\ModuleFileCredit'    => 'system/modules/filecredits/modules/ModuleFileCredit.php',
 
-	// Classes
-	'HeimrichHannot\FileCredit\Automator'           => 'system/modules/filecredits/classes/Automator.php',
-	'HeimrichHannot\FileCredit\Hooks'               => 'system/modules/filecredits/classes/Hooks.php',
-	'HeimrichHannot\FileCredit\Backend\FileCredit'  => 'system/modules/filecredits/classes/Backend/FileCredit.php',
-	'HeimrichHannot\FileCredit\Validator'           => 'system/modules/filecredits/classes/Validator.php',
-	'Contao\Feed'                                   => 'system/modules/filecredits/classes/Feed.php',
-	'HeimrichHannot\FileCredit\FileCreditIndex'     => 'system/modules/filecredits/classes/FileCreditIndex.php',
-	'HeimrichHannot\FileCredit\FileCredit'          => 'system/modules/filecredits/classes/FileCredit.php',]
+        // Classes
+        'HeimrichHannot\FileCredit\Automator'           => 'system/modules/filecredits/classes/Automator.php',
+        'HeimrichHannot\FileCredit\Hooks'               => 'system/modules/filecredits/classes/Hooks.php',
+        'HeimrichHannot\FileCredit\Backend\FileCredit'  => 'system/modules/filecredits/classes/Backend/FileCredit.php',
+        'HeimrichHannot\FileCredit\Validator'           => 'system/modules/filecredits/classes/Validator.php',
+        'Contao\Feed'                                   => 'system/modules/filecredits/classes/Feed.php',
+        'HeimrichHannot\FileCredit\FileCreditIndex'     => 'system/modules/filecredits/classes/FileCreditIndex.php',
+        'HeimrichHannot\FileCredit\FileCredit'          => 'system/modules/filecredits/classes/FileCredit.php',
+    ]
 );
 
 
@@ -47,12 +49,15 @@ ClassLoader::addClasses(
  */
 TemplateLoader::addFiles(
     [
-	'mod_filecredit_empty'                   => 'system/modules/filecredits/templates/modules',
-	'mod_filecredit'                         => 'system/modules/filecredits/templates/modules',
-	'filecredit_grouped'                     => 'system/modules/filecredits/templates/credit',
-	'filecreditgroup_copyright'              => 'system/modules/filecredits/templates/credit',
-	'filecredit_default'                     => 'system/modules/filecredits/templates/credit',
-	'be_filecredits_sync'                    => 'system/modules/filecredits/templates/backend',
-	'be_filecredits_sync_pageselection'      => 'system/modules/filecredits/templates/backend',
-	'be_filecredits_sync_pageselection_tree' => 'system/modules/filecredits/templates/backend',]
+        'mod_filecredit_empty'                   => 'system/modules/filecredits/templates/modules',
+        'mod_filecredit'                         => 'system/modules/filecredits/templates/modules',
+        'filecredit_grouped'                     => 'system/modules/filecredits/templates/credit',
+        'filecreditgroup_copyright'              => 'system/modules/filecredits/templates/credit',
+        'filecredit_default'                     => 'system/modules/filecredits/templates/credit',
+        'be_filecredits_sync'                    => 'system/modules/filecredits/templates/backend',
+        'be_filecredits_sync_pageselection'      => 'system/modules/filecredits/templates/backend',
+        'be_filecredits_sync_pageselection_tree' => 'system/modules/filecredits/templates/backend',
+        'be_filecredits_sync_pageselection_root' => 'system/modules/filecredits/templates/backend',
+    ]
+
 );